### 2018-07-01 #802 @whedon
[PRE REVIEW]: DynaML: Scala Library/REPL for Machine Learning Research 
**Submitting author:** @mandar2812 (<a href="http://orcid.org/0000-0001-6025-7113">Mandar Chandorkar</a>)
**Repository:** <a href="https://github.com/transcendent-ai-labs/DynaML" target ="_blank">https://github.com/transcendent-ai-labs/DynaML</a>
**Version:** v1.5.3-beta.2
