### 2018-07-02 #802 @mandar2812
@arfon Thanks for helping with the pdf compilation! cc @whedon

From the list of joss reviewers I see that

@rougier @nilesh-patil @xirdneh and @abuccts are interested in reviewing machine learning related submissions, although this is not an exhaustive list, I guess its a start.
