### 2018-07-02 #802 @mandar2812
@acabunoc: It would be super awesome if you want to be the editor for my manuscript on DynaML. I'm new to the JOSS process so don't really know how things work. But if you are interested, I can ask @whedon to assign you as the editor.
