### 2018-08-29 #802 @Kevin-Mattheus-Moerman
@mlgill @MikeInnes @rcurtin @rasbt @ysimillides @thomasabeel @tovbinm would you be able to review this submission for the [ _Journal of Open Source Software_ (JOSS)](http://joss.theoj.org/).
Review is focused on the [software :floppy_disk:](http://joss.theoj.org/) on a [short paper :page_facing_up: ](https://github.com/openjournals/joss-papers/blob/joss.00802/joss.00802/10.21105.joss.00802.pdf).

For more information see our [review criteria :information_source: ](https://joss.readthedocs.io/en/latest/review_criteria.html#review-criteria). Let me know if you have questions.
