### 2018-08-29 #802 @rcurtin
@Kevin-Mattheus-Moerman I'd be happy to help JOSS but I don't speak Scala (and don't have the time right now to bring myself up to speed enough in it).  Perhaps I can be more helpful for a C++ submission in the future?
