### 2018-08-29 #802 @rasbt
@Kevin-Mattheus-Moerman Unfortunately, I have to pass, too. My Scala skills are practically non-existent.
