### 2018-08-29 #802 @mandar2812
@Kevin-Mattheus-Moerman : Amongst the reviewers you proposed @tovbinm works with Scala (from his GitHub profile).

As a start, I propose @eaplatanios @sbrunk @lucataglia @aishfenton @maxpumperla @mbhushan
I can look for more members in the Scala ML community, if needed.
