### 2018-08-30 #802 @thomasabeel
I can review. This is my first experience with a review process through git, so some instructions would be welcome.
