### 2018-08-30 #802 @Kevin-Mattheus-Moerman
Thanks for your reply @tovbinm @thomasabeel. This is a pre-review issue. Once I gather enough reviewers I will initiate a review issue where the actual review will take place. At the top of a review issue each reviewer will have a simple list of checkboxes which will guide you through the review process.

* Here is an example of a completed review process: https://github.com/openjournals/joss-reviews/issues/670

* You can also check out our review guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#the-joss-review-process

Let me know if you have any other questions.
