### 2018-08-30 #802 @Kevin-Mattheus-Moerman
@tovbinm the review focuses on these items: https://joss.readthedocs.io/en/latest/review_criteria.html#review-items which form the basis of the checkboxes as well. Let me know if you can help. Thanks.
