### 2018-08-31 #802 @hclent
@Kevin-Mattheus-Moerman I am happy to help review! I do have a background with Scala as well as machine learning, so its a good fit :)
