### 2018-08-31 #802 @tovbinm
I can help. NP. (Actually I will consider submitting my own freshly open sourced project https://transmogrif.ai/)
