### 2018-09-01 #922 @whedon
[REVIEW]: DynaML: Scala Library/REPL for Machine Learning Research 
**Submitting author:** @mandar2812 (<a href="http://orcid.org/0000-0001-6025-7113">Mandar Chandorkar</a>)
**Repository:** <a href="https://github.com/transcendent-ai-labs/DynaML" target ="_blank">https://github.com/transcendent-ai-labs/DynaML</a>
**Version:** v1.5.3-beta.2

@tovbinm, @hclent, @thomasabeel, please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:

- [x] **Version:** Does the release version given match the GitHub release (v1.5.3-beta.2)?

---

- [x] **Version:** Does the release version given match the GitHub release (v1.5.3-beta.2)?

---

- [ ] **Version:** Does the release version given match the GitHub release (v1.5.3-beta.2)?
