### 2018-09-02 #922 @tovbinm
The [List of Contributors](https://transcendent-ai-labs.github.io/DynaML/about/#contributors) is missing the `sisirkoppaka` and `gitter-badger`.
