### 2018-09-02 #922 @tovbinm
**RE: Example usage**: Most examples of usage do not contain the actual data points, but rather only specify the type `val data: Stream[(DenseVector[Double], Double)] = _`. It could be useful to introduce a convenient object to allow retrieving some sample data from, e.g
```scala
val data: Stream[(DenseVector[Double], Double)] = Samples.Streams.stockPrices
```
