### 2018-09-02 #922 @tovbinm
**RE: Automated Tests:** - I looked into TravisCI builds and I could not find any tests being executed there, since the command is `sbt clean stage`, which clearly only runs the build without tests. I also noticed that nothing is reported into Codecov (https://codecov.io/gh/transcendent-ai-labs/DynaML), which I think should be fixed.
