### 2018-09-02 #922 @tovbinm
Minor: Contributing and Code of Conduct are not linked from https://transcendent-ai-labs.github.io/DynaML, but only from the README.
