### 2018-09-02 #922 @Kevin-Mattheus-Moerman
Thanks @tovbinm for that speedy review! :rocket: I'll ping you once @mandar2812 has worked on the issues you've raised.
