### 2018-10-03 #922 @hclent
@Kevin-Mattheus-Moerman I have gone through all of the bullet points, and checked off as many as I could. There are a few places where I do not believe the criteria has been fully met, or at least it needs a bit more work before publication, which I will describe below:

### Comments

* General checks and functionality were all fine
* Documentation needs improvement
    * Motivation, installation, and the examples were all good.
    * I checked off `functionality documentation` because there is decent documentation in `docs` **BUT** I think that the actual source code needs link back to these. For example, [GeneralizedLinearModel.scala](https://github.com/transcendent-ai-labs/DynaML/blob/8a565e1da39e2c28294a976becf837d40ef2c7d1/dynaml-core/src/main/scala-2.11/io/github/mandar2812/dynaml/models/lm/GeneralizedLinearModel.scala) should have a link back to [the documentation](https://transcendent-ai-labs.github.io/DynaML/core/core_glm/), and the documentation should have a link back to the class. The code and the docs are oddly disconnected, which means that an unnecessary amount of digging around has to be done by the users to get back to the source code, in my opinion.
    * So I can see that the [TRAVIS build](https://travis-ci.org/transcendent-ai-labs/DynaML) is passing, but looking at the tests, I think there needs to be more documentation. I could not find any documentation about _which_ tests in `dynaml-core/src/test` that a potential contributor needs to update if they are going to contribute or about how the testing process plays into contribution at all.
    * Some tests seem incomplete, for example [WaveletSpec.scala](https://github.com/transcendent-ai-labs/DynaML/blob/master/dynaml-core/src/test/scala-2.11/io/github/mandar2812/dynaml/wavelets/WaveletSpec.scala)
* Per the statement of need in the software paper, the author(s) describe what `dynaml` gives (e.g. its end-to-end), but I don't think the need for anything is explained _explicitly_, and it should be! A good way to do this could be by citing other ML softwares that work on the JVM, and what `dynaml` has that they don't.

All in all, this is a really cool project and I think that a few additional updates to documentation and the paper will make it suitable for publication :)
