### 2018-11-20 #922 @arfon
> @arfon how do we assign a "pause" label here?

Just use the label picker on the top right of the issue ☝️
