### 2019-05-11 #922 @labarba
@mandar2812 last [commented on Nov 20, 2018](https://github.com/openjournals/joss-reviews/issues/922#issuecomment-440215126) that time was scarce to work on the improvements requested during review. It has now been nearly 6 months, and we should either have a commitment from the author that they will work on this revision, or withdraw the submission.
